+++ src/vars/commandData.ts
@@ -6,12 +6,12 @@ const mods = 'See [here](https://sbrstrkkdwmdr.github.io/projects/ssob_docs/comm
 const scoreListString =
     `Mods can be specified with +[mods], -mx [exact mods] or -me [exclude mods]
 The arguments \`pp\`, \`score\`, \`acc\`, \`bpm\` and \`miss\` use the following format:
-\`-key value\` to filter by that exact value (ie. -bpm 220)
-\`-key >value\` to filter scores below value (ie. -pp >500)
-\`-key <value\` to filter scores above value (ie. -acc <90)
-\`-key min..max\` to filter scores between min and max (ie. -miss 1..3)
+\`-key value\` to filter by that exact value (eg. -bpm 220)
+\`-key >value\` to filter scores below value (eg. -pp >500)
+\`-key <value\` to filter scores above value (eg. -acc <90)
+\`-key min..max\` to filter scores between min and max (eg. -miss 1..3)
 The \`sort\` arg can be specified using -value (ie -recent)
-You can also show a single score by using \`-parse <id>\` (ie. -parse 5)
+You can also show a single score by using \`-parse <id>\` (eg. -parse 5)
 `;
 
 /**